feat/refactor( ProductService.kt, ProductController.kt ): create Service layer between ProductController and ProductRepository
